added messages api routes to message view config

+++ app/backend/messages/views.py
@@ -11,31 +11,25 @@ from messages.api.delete_request import delete_messages_request
 
 
 class Messages(View):
-    curr_user = User()
     # Define the function that will handle GET requests
-
     async def get(self, request: HttpRequest):
         res = await get_messages_request(request)
         # Return a simple HttpResponse
         return HttpResponse(res)
 
     async def post(self, request: HttpRequest):
-        signed_in = await self.curr_user.signed_in(request)
+        curr_user = User()
+        signed_in = await curr_user.signed_in(request)
         if signed_in is None:
             return HttpResponse('User not signed in', status=401)
-        res = await post_messages_request(request, user_id=signed_in)
-        return HttpResponse(res)
-
-    async def put(self, request: HttpRequest):
-        signed_in = await self.curr_user.signed_in(request)
-        if signed_in is None:
-            return HttpResponse('User not signed in', status=401)
-        res = await put_messages_request(request, user_id=signed_in)
-        return HttpResponse(res)
-
-    async def delete(self, request: HttpRequest):
-        signed_in = await self.curr_user.signed_in(request)
-        if signed_in is None:
-            return HttpResponse('User not signed in', status=401)
-        res = await delete_messages_request(request, user_id=signed_in)
+        # we will use the _method to determine the type of request
+        # since django cannot template delete and put requests
+        # so we will default to post request entirely
+        method = self.request.POST.get('_method', '').lower()
+        if method == 'put':
+            res = await put_messages_request(request, user_id=signed_in)
+        if method == 'delete':
+            res = await delete_messages_request(request, user_id=signed_in)
+        if method == 'post':
+            res = await post_messages_request(request, user_id=signed_in)
         return HttpResponse(res)
